complete Entity::module Factory::pattern and Prototype::pattern
