+++ app/core/Modules/Sitemap/Sitemap.php
@@ -3,8 +3,7 @@
 namespace Blog\Modules\Sitemap;
 
 use Blog\Modules\DateFormat\DateFormat;
-use Blog\Modules\Entity\ArticlePrototype;
-use Blog\Modules\View\Blog;
+use Blog\Modules\Entity\Article;
 
 class Sitemap
 {
@@ -28,13 +27,13 @@ class Sitemap
         $update = \Blog\Modules\View\Blog::lastUpdate();
         $links['/']['lastmod'] = $links['/blog']['lastmod'] = new DateFormat($update, DateFormat::COMPLETE);
         // generate blog articles data
-        foreach (ArticlePrototype::loadList(['load_with_comments' => false]) as $article) {
+        foreach (Article::loadList(['load_with_comments' => false]) as $article) {
             if ($article->exists()) {
                 $url = $article->url();
                 $links[$url] = [
                     'loc' => fullUrlTo($url),
-                    'priority' => ArticlePrototype::getSitemapPriority(),
-                    'changefreq' => ArticlePrototype::getSitemapChangefreq(),
+                    'priority' => Article::getSitemapPriority(),
+                    'changefreq' => Article::getSitemapChangefreq(),
                     'lastmod' => new DateFormat($article->get('updated'), DateFormat::COMPLETE)
                 ];
             }
